# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Backend mandiri Tuléh POS. Amplop respons: {success, data, meta, message, errors}.",
        "title": "Tuléh Server API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/gateway/midtrans": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gateway"
                ],
                "summary": "Konfigurasi Midtrans merchant (server key BERTOPENG)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/respond.Amplop"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.GatewayResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Ditolak (403) bila modul Midtrans belum diaktifkan platform. server_key dikosongkan → key tersimpan dipertahankan.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gateway"
                ],
                "summary": "Simpan konfigurasi Midtrans (Owner/Manager)",
                "parameters": [
                    {
                        "description": "Konfigurasi",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.SimpanGatewayRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/respond.Amplop"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.GatewayResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "403": {
                        "description": "Modul Midtrans belum aktif",
                        "schema": {
                            "$ref": "#/definitions/respond.Amplop"
                        }
                    }
                }
            }
        },
        "/gateway/midtrans/status": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "siap=true → tampilkan opsi QRIS dinamis; client_key aman dipakai SDK klien.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gateway"
                ],
                "summary": "Status Midtrans untuk aplikasi kasir (tanpa server key)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/respond.Amplop"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.StatusGatewayResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
        "/usahas": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "usaha"
                ],
                "summary": "Daftar usaha/merchant (SUPERADMIN)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Cari nama/kode",
                        "name": "q",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Halaman",
                        "name": "page",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/respond.Amplop"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/http.UsahaResponse"
                                            }
                                        },
                                        "meta": {
                                            "$ref": "#/definitions/respond.Meta"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Owner langsung bisa login dan mengelola usahanya (produk, kasir, dll).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "usaha"
                ],
                "summary": "Buat usaha BARU + akun owner-nya sekaligus (SUPERADMIN)",
                "parameters": [
                    {
                        "description": "Usaha + owner",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.BuatUsahaRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/respond.Amplop"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.UsahaResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "409": {
                        "description": "Email owner / kode usaha sudah dipakai",
                        "schema": {
                            "$ref": "#/definitions/respond.Amplop"
                        }
                    }
                }
            }
        },
        "/usahas/{id}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "usaha"
                ],
                "summary": "Ubah nama / SUSPEND usaha (aktif=false menolak login penggunanya)",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID usaha",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Perubahan",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/http.UbahUsahaRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/respond.Amplop"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/http.UsahaResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "http.BuatUsahaRequest": {
            "type": "object",
            "required": [
                "nama",
                "owner_email",
                "owner_nama",
                "owner_password"
            ],
            "properties": {
                "kode": {
                    "type": "string",
                    "maxLength": 30
                },
                "nama": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 2
                },
                "owner_email": {
                    "type": "string",
                    "maxLength": 150
                },
                "owner_nama": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 2
                },
                "owner_password": {
                    "type": "string",
                    "maxLength": 72,
                    "minLength": 8
                }
            }
        },
...
        "http.GatewayResponse": {
            "type": "object",
            "properties": {
                "aktif": {
                    "type": "boolean"
                },
                "client_key": {
                    "type": "string"
                },
                "env": {
                    "type": "string"
                },
                "merchant_id": {
                    "type": "string"
                },
                "platform_aktif": {
                    "type": "boolean"
                },
                "server_key_hint": {
                    "type": "string"
                },
                "server_key_terisi": {
                    "type": "boolean"
                },
                "siap": {
                    "type": "boolean"
                }
            }
        },
...
        "http.SimpanGatewayRequest": {
            "type": "object",
            "required": [
                "aktif",
                "env"
            ],
            "properties": {
                "aktif": {
                    "type": "boolean"
                },
                "client_key": {
                    "type": "string",
                    "maxLength": 120
                },
                "env": {
                    "type": "string",
                    "enum": [
                        "sandbox",
                        "production"
                    ]
                },
                "merchant_id": {
                    "type": "string",
                    "maxLength": 60
                },
                "server_key": {
                    "description": "kosong = tak diubah",
                    "type": "string",
                    "maxLength": 120
                }
            }
        },
...
        "http.StatusGatewayResponse": {
            "type": "object",
            "properties": {
                "client_key": {
                    "type": "string"
                },
                "env": {
                    "type": "string"
                },
                "siap": {
                    "type": "boolean"
                }
            }
        },
...
        "http.UbahUsahaRequest": {
            "type": "object",
            "properties": {
                "aktif": {
                    "type": "boolean"
                },
                "midtrans_aktif": {
                    "type": "boolean"
                },
                "nama": {
                    "type": "string",
                    "maxLength": 150,
                    "minLength": 2
                }
            }
        },
        "http.UsahaResponse": {
            "type": "object",
            "properties": {
                "aktif": {
                    "type": "boolean"
                },
                "dibuat": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "kode": {
                    "type": "string"
                },
                "midtrans_aktif": {
                    "type": "boolean"
                },
                "nama": {
                    "type": "string"
                }
            }
        },
...
        "respond.Amplop": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/respond.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "respond.Meta": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "last_page": {
                    "type": "integer"
                },
                "per_page": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Isi dengan: Bearer \u003caccess_token\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}